fixes to page controller along with permissions

+++ src/main/java/six/team/backend/controller/PagesController.java
@@ -7,6 +7,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import six.team.backend.PageJsonGen;
+import six.team.backend.dao.UserDAO;
 import six.team.backend.model.News;
 import six.team.backend.model.Pages;
 import six.team.backend.store.CommentStore;
@@ -26,41 +27,69 @@ public class PagesController {
 
     @RequestMapping(produces = MediaType.APPLICATION_JSON_VALUE, method = RequestMethod.POST)
     public@ResponseBody ResponseEntity<String> addPage(HttpServletRequest request, HttpServletResponse response, @RequestBody String text) {
-        boolean exists= Pages.checkValidity(request.getHeader("title"));
-        JSONObject json = new JSONObject();
-        json.put("exists", exists);
+        UserDAO UD = new UserDAO();
+        String token = request.getHeader("token");
+        if(UD.getUserGroupPermissions(UD.getUserGroup(token),"pageadd")) {
+            boolean exists = Pages.checkValidity(request.getHeader("title"));
+            JSONObject json = new JSONObject();
+            json.put("exists", exists);
 
-        if(exists){
-            return new ResponseEntity<String>(json.toString(), HttpStatus.valueOf(409));
-        }
-        else{
-            boolean success=Pages.addPage(request.getHeader("parentSlug"),request.getHeader("title"),request.getHeader("description"),text,request.getHeader("permission"),Integer.parseInt(request.getHeader("order")));
-            json.put("sucess", success);
-            if(success)
-                return new ResponseEntity<String>(json.toString(), HttpStatus.valueOf(201));
-            else
-                return new ResponseEntity<String>(json.toString(), HttpStatus.valueOf(201));
+            if (exists) {
+                return new ResponseEntity<String>(json.toString(), HttpStatus.valueOf(409));
+            } else {
+                boolean success = Pages.addPage(request.getHeader("parentSlug"), request.getHeader("title"), request.getHeader("description"), text, request.getHeader("permission"), Integer.parseInt(request.getHeader("order")));
+                json.put("sucess", success);
+                if (success)
+                    return new ResponseEntity<String>(json.toString(), HttpStatus.valueOf(201));
+                else
+                    return new ResponseEntity<String>(json.toString(), HttpStatus.valueOf(201));
+            }
+        }else {
+            JSONObject message = new JSONObject();
+            message.put("pages", "You are unauthorized to add a page");
+            return new ResponseEntity<String>(message.toString(), HttpStatus.UNAUTHORIZED);
         }
 
     }
 
     @RequestMapping(produces = MediaType.APPLICATION_JSON_VALUE,value = "/{slug}/delete", method = RequestMethod.POST)
     public@ResponseBody ResponseEntity<String> deletePage(HttpServletRequest request, HttpServletResponse response, @PathVariable(value="slug") String slug) {
-        boolean success=Pages.deletePage(slug);
-        if(success)
-            return new ResponseEntity<String>("Page deleted", HttpStatus.valueOf(201));
-        else
-            return new ResponseEntity<String>("Page couldnt be deleted", HttpStatus.valueOf(201));
+        UserDAO UD = new UserDAO();
+        String token = request.getHeader("token");
+        if(UD.getUserGroupPermissions(UD.getUserGroup(token),"pagedelete")) {
+            boolean success = Pages.deletePage(slug);
+            JSONObject json = new JSONObject();
+            json.put("sucess", success);
+            if (success) {
+                return new ResponseEntity<String>(json.toString(), HttpStatus.valueOf(201));
+            } else {
+                return new ResponseEntity<String>(json.toString(), HttpStatus.valueOf(201));
+            }
+        }else {
+            JSONObject message = new JSONObject();
+            message.put("pages", "You are unauthorized to delete a page");
+            return new ResponseEntity<String>(message.toString(), HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @RequestMapping(produces = MediaType.APPLICATION_JSON_VALUE,value = "/{slug}/edit", method = RequestMethod.POST)
     public@ResponseBody ResponseEntity<String> editPage(HttpServletRequest request, HttpServletResponse response,@PathVariable(value="slug") String slug, @RequestBody String text) {
-        boolean success=Pages.editPage(slug,request.getHeader("parentSlug"), request.getHeader("title"), request.getHeader("description"), text, request.getHeader("permission"), Integer.parseInt(request.getHeader("order")));
-        if(success)
-            return new ResponseEntity<String>("Page edited", HttpStatus.valueOf(201));
-        else
-            return new ResponseEntity<String>("Page couldn be edited", HttpStatus.valueOf(201));
-
+        UserDAO UD = new UserDAO();
+        String token = request.getHeader("token");
+        if(UD.getUserGroupPermissions(UD.getUserGroup(token),"pagedelete")) {
+            boolean success = Pages.editPage(slug, request.getHeader("parentSlug"), request.getHeader("title"), request.getHeader("description"), text, request.getHeader("permission"), Integer.parseInt(request.getHeader("order")));
+            JSONObject json = new JSONObject();
+            json.put("sucess", success);
+            if (success) {
+                return new ResponseEntity<String>(json.toString(), HttpStatus.valueOf(201));
+            } else {
+                return new ResponseEntity<String>(json.toString(), HttpStatus.valueOf(201));
+            }
+        }else {
+            JSONObject message = new JSONObject();
+            message.put("pages", "You are unauthorized to edit a page");
+            return new ResponseEntity<String>(message.toString(), HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @RequestMapping(produces = MediaType.APPLICATION_JSON_VALUE, method = RequestMethod.GET)
